+++ lib/src/query/public_group_channel_list_query.dart
@@ -8,16 +8,16 @@ import '../sdk/sendbird_sdk_api.dart';
 
 /// A query object to retrieve list of public group channel.
 class PublicGroupChannelListQuery extends QueryBase {
-  /// Filter for channel urls.
+  /// Filters for channel urls.
   ///
   /// Result will return a list containing only and exactly matched
   /// with given urls. This filter does not cooperate with other filters.
   List<String> channelUrls;
 
-  /// Filter channels for super group channel. Default is `all`
+  /// Filters channels for super group channel. Default is `all`
   GroupChannelSuperChannelFilter superChannelFilter;
 
-  /// Filter by current user's member state in public group channel
+  /// Filters by current user's member state in public group channel
   PublicGroupChannelMembershipFilter membershipFilter;
 
   /// Order of query result
@@ -26,10 +26,10 @@ class PublicGroupChannelListQuery extends QueryBase {
   /// Filter by custom type that starts with
   String customTypeStartWithFilter;
 
-  /// Filter by custom types.
+  /// Filters by custom types.
   List<String> customTypesFilter;
 
-  /// Filter by channel name contains certain text
+  /// Filters by channel name contains certain text
   String channelNameContainsFilter;
 
   /// Sets metadata's key for ordering its value