+++ lib/src/utils/logger.dart
@@ -13,6 +13,21 @@ var logger = Logger(
   ),
 );
 
+void initLogger(LogLevel lvl) {
+  final level = convertLogLevel(lvl);
+  logger = Logger(
+    level: level,
+    printer: PrettyPrinter(
+      methodCount: 0, // number of method calls to be displayed
+      errorMethodCount: 8, // number of method calls if stacktrace is provided
+      lineLength: 150, // width of the output
+      colors: true, // Colorful log messages
+      printEmojis: true, // Print an emoji for each log message
+      printTime: false, // Should each log print contain a timestamp
+    ),
+  );
+}
+
 Level convertLogLevel(LogLevel level) {
   switch (level) {
     case LogLevel.debug: